requirements.txt
Flask == 1.1.2
Flask-Script == 2.0.6
Flask-SQLAlchemy == 2.4.4
gunicorn == 20.0.4
Jinja2 == 2.11.2
line-bot-sdk == 1.18.0
numpy ==  1.18.1
requests == 2.23.0
requests-oauthlib == 1.3.0
SQLAlchemy == 1.3.18
urllib3 == 1.25.9
